[board] Move board support defines into src/boards files
Also, add src/boards/README.md to clarify the intention
